Added eslint for editors, refactored components and added stylesheets
