+++ package.json
@@ -8,6 +8,11 @@
     "react-router-dom": "^4.0.0"
   },
   "devDependencies": {
+    "eslint": "^3.15.0",
+    "eslint-config-airbnb": "^14.1.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.9.0",
     "react-scripts": "0.9.5"
   },
   "scripts": {
