+++ src/App.css
@@ -2,11 +2,6 @@
   text-align: center;
 }
 
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 80px;
-}
-
 .App-header {
   background-color: #222;
   height: 150px;
@@ -14,6 +9,7 @@
   color: white;
 }
 
-.App-intro {
-  font-size: large;
+.App-header h2 a {
+  color: white;
+  text-decoration: none;
 }
