+++ src/Components/Navigation/NavigationBar.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-const NavigationBar = () => {
+import './NavigationBar.css';
+
+function NavigationBar() {
   return (
-    <ul>
+    <ul className="navigationBar">
       <li>
         <Link to="/faq">
           Faq
